deleting files on history_audio now working

+++ delete_query.php
@@ -14,25 +14,24 @@ if($_POST['clearAll'] == 'true'){
         $deleteQuery = mysqli_prepare($dbcon, "DELETE FROM audio_files WHERE user_id = ?");
         bindAndExec($deleteQuery, "s", $userId);
 
-        // since we delete all, get all the files that has the user_id in file and delete them
-        $mask = $userId . "_*";
+        // since we delete all, get all the files and folder that has the user_id in file
+        $mask = "audio_files/" . $userId . "_*";
 
-        $iterator = new RecursiveIteratorIterator(
-                new RecursiveDirectoryIterator($mask, FilesystemIterator::SKIP_DOTS),
-                RecursiveIteratorIterator::CHILD_FIRST
-            );
+        // put all the folder/filenames that matches the $mask in an array
+        $items = glob($mask);
 
-        //$items = glob($mask);
-
-        foreach ($iterator as $file) {
-            if ($file->isFile()) {
-                unlink($file->getPathname());
-            } elseif ($file->isDir() && !$iterator->hasChildren()) {
-                rmdir($file->getPathname());rmdir($item);
+        
+        foreach ($items as $file) {
+        
+            // check if it's an audio file or a folder, then delete them 
+            //  using their designated step
+            if (is_file($file)) {
+                unlink($file);
+            } elseif (is_dir($file)) {
+                removeFolder($file);
             }
+        
         }
-
-        //array_map('unlink', glob($mask));
     }
     else{            
         // Deletes all rows corresponding to user from text_translation table
@@ -42,25 +41,40 @@ if($_POST['clearAll'] == 'true'){
  }
 
  elseif($_POST['deleteRows'] == 'true'){
+
+    $userId = $_POST['userId'];
+
     if($_POST['fromAudio'] == 1){
         $rowsToDelete = json_decode($_POST['rowsToDelete']);
         $filesToDelete = json_decode($_POST['filesToDelete']);
         for($i = 0; $i < count($rowsToDelete); $i++){
                 $deleteRows = $rowsToDelete[$i];
                 $deleteFiles =$filesToDelete[$i];
+
+                $audiofile = fetchAudioContent($deleteFiles, $userId);
+                
+                $filenameMask = "audio_files/" . $userId . "_" . $audiofile['filename'] . $audiofile['formatted_date'] . "." . $audiofile['extension']; 
+                if (file_exists($filenameMask)) 
+                    unlink($filenameMask);
+        
+                $folderMask = "audio_files/" . $userId . "_" . $audiofile['filename'] . $audiofile['formatted_date'];
+                removeFolder($folderMask);
+                
                 $deleteQuery = mysqli_prepare($dbcon, "DELETE FROM text_translations WHERE text_id = ?");
                 bindAndExec($deleteQuery, "s", $deleteRows);
 
                 $deleteQuery = mysqli_prepare($dbcon, "DELETE FROM audio_files WHERE file_id = ?");
                 bindAndExec($deleteQuery, "s", $deleteFiles);
 
+                
             }
     }
 
     else{
         foreach(json_decode($_POST['rowsToDelete']) as $rowNum){
             $deleteQuery = mysqli_prepare($dbcon, "DELETE FROM text_translations WHERE text_id = ?");
-            bindAndExec($deleteQuery, "s", $rowNum);        }
+            bindAndExec($deleteQuery, "s", $rowNum);        
+        }
     }
     
  }
@@ -77,33 +91,20 @@ if($_POST['clearAll'] == 'true'){
 
     //deletes audio file record from database if it's an audio to text translation
     if($_POST['fileId'] != null){
-        $datequery = "SELECT file_name, DATE_FORMAT(upload_date, '%m%d%Y_%H%i%s') AS formatted_date 
-                        FROM audio_files WHERE file_id = '$fileId'";
-        $dateresult = mysqli_query($dbcon, $datequery);
-        $row = mysqli_fetch_assoc($dateresult);
+        
+        //get the audio file content from the database, to be used on locating the filename in audio_files folder
+        $audiofile = fetchAudioContent($fileId, $userId);
 
-        $filename = pathinfo($row['file_name'], PATHINFO_FILENAME);
-        $extension = pathinfo($row['file_name'], PATHINFO_EXTENSION);
-
-        $filenameMask = "audio_files/" . $userId . "_" . $filename . $row['formatted_date'] . $extension;
-        unlink($filenameMask);
-
-        /*$iterator = new RecursiveIteratorIterator(
-            new RecursiveDirectoryIterator($directoryMask, FilesystemIterator::SKIP_DOTS),
-            RecursiveIteratorIterator::CHILD_FIRST
-        );
-     
-        foreach ($iterator as $item) {
-            if ($item->isFile()) {
-                unlink($item->getPathname());
-            } elseif ($item->isDir() && !$iterator->hasChildren()) {
-                rmdir($item->getPathname());
-            }
-        }*/
+        // remove the audio file uploaded by user
+        $filenameMask = "audio_files/" . $userId . "_" . $audiofile['filename'] . $audiofile['formatted_date'] . "." . $audiofile['extension']; 
+        if (file_exists($filenameMask)) 
+            unlink($filenameMask);
+        
+        // remove the folder created on specific file and all its contents
+        $folderMask = "audio_files/" . $userId . "_" . $audiofile['filename'] . $audiofile['formatted_date'];
+        removeFolder($folderMask);
         
 
-        
-        
         
         $deleteQuery = mysqli_prepare($dbcon, "DELETE FROM audio_files WHERE file_id = ?");
         bindAndExec($deleteQuery, "s", $fileId); 
@@ -127,4 +128,47 @@ exit(json_encode($result));
 function bindAndExec($stmt, $markers, $value){
     mysqli_stmt_bind_param($stmt, $markers, $value);
     mysqli_stmt_execute($stmt);
+}
+
+
+
+function removeFolder($path) {
+    // process of deleting the folder in audio files
+    if (is_dir($path)) {
+        $iterator = new DirectoryIterator($path);
+    
+        
+        // loop through each content inside the folder
+        foreach($iterator as $file) {
+            if ($file->isFile()) {
+                unlink($file->getRealPath()); // get the relative path and delete the file
+            } elseif (!$file->isDot() && $file->isDir()) {
+                rmdir($file->getRealPath()); // get the relative path and delete the folder
+            }
+        }
+        
+        // finally, delete the now-empty folder
+        rmdir($path);
+    }
+}
+
+function fetchAudioContent($fileid, $userid) {
+    global $dbcon;
+    $datequery = "SELECT file_name, DATE_FORMAT(upload_date, '%m%d%Y_%H%i%s') AS formatted_date 
+                        FROM audio_files WHERE file_id = '$fileid'";
+    $dateresult = mysqli_query($dbcon, $datequery);
+
+    $audiofile =  mysqli_fetch_assoc($dateresult);
+
+    // get the filename and extension
+    $filename = pathinfo($audiofile['file_name'], PATHINFO_FILENAME);
+    $extension = pathinfo($audiofile['file_name'], PATHINFO_EXTENSION);
+
+    // put filename, extension and date in array, these infos are needed to select
+    //  specific record from the database in order to delete a specific file
+    $info = array("filename" => $filename, 
+                "extension" => $extension,
+                "formatted_date" => $audiofile['formatted_date']);
+
+    return $info;
 }
